Extract shared export-utils from docx-actions
Move the image fetcher/uploader, downloadBlob, pickFile, and
safeFilename helpers out of docx-actions.ts into a new
export-utils.ts so the upcoming PDF export action can reuse them.

docx-actions.ts now keeps only the DOCX-specific import/export
wrappers and re-exports docxImageUploader/docxImageFetcher as
back-compat aliases for existing callers.
